refactor: rename error, schema, middleware and type files
