+++ src/modules/auth/auth.service.ts
@@ -2,8 +2,8 @@ import { LoginRequestBodyDto } from "./dto/request/login-request-body.dto";
 import bcrypt from "bcrypt";
 import { SignUpRequestBodyDto } from "./dto/request/sign-up-request-body.dto";
 import { UsersRepository } from "../users/users.repository";
-import { BadRequestError } from "#/shared/errors/bad-request-error";
-import { UnauthorizedError } from "#/shared/errors/unauthorized-error";
+import { BadRequestError } from "#/shared/errors/bad-request.error";
+import { UnauthorizedError } from "#/shared/errors/unauthorized.error";
 import { generateJwtToken } from "#/shared/utils/generateToken";
 import { User } from "../users/entities/user.entity";
 import { envConfig } from "#/config/env.config";
