+++ src/modules/comments/comments.controller.ts
@@ -2,9 +2,9 @@ import "express-async-errors";
 import { Request, Response, Router } from "express";
 import { CommentsService } from "./comments.service";
 import { Comment } from "./entities/comment.entity";
-import { checkResourceOwnership } from "#/shared/middlewares/checkResourceOwnership.middleware";
+import { checkResourceOwnership } from "#/shared/middlewares/check-resource-ownership.middleware";
 import { validatePathParameter } from "#/shared/validators/path-parameter.validator";
-import { uuidSchema } from "#/shared/schemas/uuid-schema";
+import { uuidSchema } from "#/shared/schemas/uuid.schema";
 import { auth } from "#/shared/middlewares/auth.middleware";
 import { validateRequestBody } from "#/shared/validators/request-body.validator";
 import { updateCommentRequestBodyDtoSchema } from "./dto/request/update-comment-request-body.dto";
