+++ src/shared/errors/bad-request.error.ts
@@ -1,4 +1,4 @@
-import { CustomError } from "#/shared/errors/custom-error";
+import { CustomError } from "#/shared/errors/custom.error";
 
 export class BadRequestError extends CustomError {
 	constructor(message: string) {
