+++ src/shared/errors/custom.error.ts
