+++ src/shared/errors/not-found.error.ts
@@ -1,4 +1,4 @@
-import { CustomError } from "#/shared/errors/custom-error";
+import { CustomError } from "#/shared/errors/custom.error";
 
 export class NotFoundError extends CustomError {
 	constructor(message: string) {
