+++ src/shared/errors/unauthorized.error.ts
@@ -1,4 +1,4 @@
-import { CustomError } from "#/shared/errors/custom-error";
+import { CustomError } from "#/shared/errors/custom.error";
 
 export class UnauthorizedError extends CustomError {
 	constructor(message: string) {
