+++ src/shared/middlewares/not-found.middleware.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
 
-export function notFoundMiddleware(req: Request, res: Response) {
+export function notFoundMiddleware(_req: Request, res: Response) {
 	res.status(404).json({
 		message: "Not Found",
 	});
