+++ src/shared/utils/verifyToken.ts
@@ -1,7 +1,7 @@
 import { jwtSecret } from "#/config/jwt-secret.config";
 import jwt, { JsonWebTokenError } from "jsonwebtoken";
-import { AuthTokenPayload } from "../types/jwtPayload";
-import { UnauthorizedError } from "../errors/unauthorized-error";
+import { AuthTokenPayload } from "../types/jwt-payload.types";
+import { UnauthorizedError } from "../errors/unauthorized.error";
 
 export function verifyToken(token: string): AuthTokenPayload | null {
 	try {
