+++ tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.ts","./src/main.ts","./src/config/dataSource.config.ts","./src/config/jwt-secret.config.ts","./src/database/database.ts","./src/database/migrations/1728030204605-create-users.ts","./src/modules/auth/auth.controller.ts","./src/modules/auth/auth.repository.ts","./src/modules/auth/auth.service.ts","./src/modules/auth/dto/request/login-request-body.dto.ts","./src/modules/auth/dto/request/sign-up-request-body.dto.ts","./src/modules/auth/tests/auth.controller.spec.ts","./src/modules/auth/tests/auth.service.spec.ts","./src/modules/users/users.model.ts","./src/modules/users/users.repository.ts","./src/shared/errors/bad-request-error.ts","./src/shared/errors/custom-error.ts","./src/shared/errors/not-found-error.ts","./src/shared/errors/unauthorized-error.ts","./src/shared/middlewares/auth.middleware.ts","./src/shared/middlewares/error.middleware.ts","./src/shared/middlewares/not-found.middleware.ts","./src/shared/types/jwtPayload.ts","./src/shared/utils/generateToken.ts","./src/shared/utils/sendCookies.ts","./src/shared/utils/verifyToken.ts","./src/shared/validators/request-body.validator.ts"],"version":"5.6.2"}
+{"root":["./src/app.ts","./src/main.ts","./src/config/dataSource.config.ts","./src/config/env.config.ts","./src/config/jwt-secret.config.ts","./src/database/database.ts","./src/database/migrations/1728406995519-create-users.ts","./src/database/migrations/1728407079841-create-blogs.ts","./src/database/migrations/1728555721876-added-blogs-cascade-delete.ts","./src/database/migrations/1728636001316-create-comments.ts","./src/modules/auth/auth.controller.ts","./src/modules/auth/auth.service.ts","./src/modules/auth/dto/request/login-request-body.dto.ts","./src/modules/auth/dto/request/sign-up-request-body.dto.ts","./src/modules/auth/tests/auth.controller.spec.ts","./src/modules/auth/tests/auth.service.spec.ts","./src/modules/blogs/blogs.controller.ts","./src/modules/blogs/blogs.repository.ts","./src/modules/blogs/blogs.service.ts","./src/modules/blogs/dto/request/blog-search-params.dto.ts","./src/modules/blogs/dto/request/create-blog-request-body.dto.ts","./src/modules/blogs/dto/request/update-blog-request-body.dto.ts","./src/modules/blogs/entities/blog.entity.ts","./src/modules/blogs/tests/blogs.controller.spec.ts","./src/modules/blogs/tests/blogs.service.spec.ts","./src/modules/comments/comments.controller.ts","./src/modules/comments/comments.repository.ts","./src/modules/comments/comments.service.ts","./src/modules/comments/dto/request/comment-search-params.dto.ts","./src/modules/comments/dto/request/create-comment-request-body.dto.ts","./src/modules/comments/dto/request/update-comment-request-body.dto.ts","./src/modules/comments/entities/comment.entity.ts","./src/modules/comments/tests/comments.controller.spec.ts","./src/modules/comments/tests/comments.service.spec.ts","./src/modules/users/users.repository.ts","./src/modules/users/entities/user.entity.ts","./src/shared/errors/bad-request.error.ts","./src/shared/errors/custom.error.ts","./src/shared/errors/not-found.error.ts","./src/shared/errors/unauthorized.error.ts","./src/shared/middlewares/auth.middleware.ts","./src/shared/middlewares/check-resource-ownership.middleware.ts","./src/shared/middlewares/error.middleware.ts","./src/shared/middlewares/not-found.middleware.ts","./src/shared/schemas/pagination-search-params.schema.ts","./src/shared/schemas/uuid.schema.ts","./src/shared/types/express.types.d.ts","./src/shared/types/jwt-payload.types.ts","./src/shared/utils/generateToken.ts","./src/shared/utils/sendCookies.ts","./src/shared/utils/verifyToken.ts","./src/shared/validators/path-parameter.validator.ts","./src/shared/validators/request-body.validator.ts","./src/shared/validators/search-params.validator.ts"],"version":"5.6.2"}
